refactor(e2e): add user activation helper and update team member tests
